+++ src/examples/java/io/nats/examples/NatsJsPullFetch.java
@@ -16,8 +16,7 @@ package io.nats.examples;
 import io.nats.client.*;
 
 import java.time.Duration;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.List;
 
 import static io.nats.examples.ExampleUtils.sleep;
 
@@ -25,49 +24,19 @@ import static io.nats.examples.ExampleUtils.sleep;
  * This example will demonstrate a pull subscription with:
  * - a batch size + fetch i.e. subscription.fetch(10, Duration.ofSeconds(10))
  *
- * Usage: java NatsJsFetch [-s server] [-strm stream] [-sub subject] [-dur durable]
+ * Usage: java NatsJsPullFetch [-s server] [-strm stream] [-sub subject] [-dur durable]
  *   Use tls:// or opentls:// to require tls, via the Default SSLContext
  *   Set the environment variable NATS_NKEY to use challenge response authentication by setting a file containing your private key.
  *   Set the environment variable NATS_CREDS to use JWT/NKey authentication by setting a file containing your user creds.
  *   Use the URL for user/pass/token authentication.
  */
-public class NatsJsReceive extends NatsJsPullSubBase {
-    static class ReceiveMessageHandler implements MessageHandler {
-        boolean ack;
-        int expected;
-        AtomicInteger received;
-        CountDownLatch latch;
-
-        public ReceiveMessageHandler(boolean ack, int expected) {
-            this.ack = ack;
-            this.expected = expected;
-            received = new AtomicInteger();
-            latch = new CountDownLatch(1);
-            System.out.print("Receive ->");
-        }
-
-        @Override
-        public void onMessage(Message msg) throws InterruptedException {
-            if (msg == null) {
-                System.out.println(" <- ");
-                System.out.println("We should have received " + expected + " total messages, we received: " + received);
-                latch.countDown();
-            }
-            else {
-                received.incrementAndGet();
-                System.out.print(" " + new String(msg.getData()));
-                if (ack) {
-                    msg.ack();
-                }
-            }
-        }
-    }
+public class NatsJsPullFetch extends NatsJsPullSubBase {
 
     public static void main(String[] args) {
         ExampleArgs exArgs = ExampleArgs.builder()
-                .defaultStream("receive-ack-stream")
-                .defaultSubject("receive-ack-subject")
-                .defaultDurable("receive-ack-durable")
+                .defaultStream("fetch-ack-stream")
+                .defaultSubject("fetch-ack-subject")
+                .defaultDurable("fetch-ack-durable")
                 .build(args);
         
         try (Connection nc = Nats.connect(ExampleUtils.createExampleOptions(exArgs.server))) {
@@ -98,73 +67,80 @@ public class NatsJsReceive extends NatsJsPullSubBase {
             // 1. Fetch, but there are no messages yet.
             // -  Read the messages, get them all (0)
             System.out.println("----------\n1. There are no messages yet");
-
-            ReceiveMessageHandler handler = new ReceiveMessageHandler(true, 0);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            List<Message> messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 0 total messages, we received: " + messages.size());
 
             // 2. Publish 10 messages
             // -  Fetch messages, get 10
             System.out.println("----------\n2. Publish 10 which satisfies the batch");
             publish(js, exArgs.subject, "A", 10);
-            handler = new ReceiveMessageHandler(true, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 10 total messages, we received: " + messages.size());
 
             // 3. Publish 20 messages
             // -  Fetch messages, only get 10
             System.out.println("----------\n3. Publish 20 which is larger than the batch size.");
             publish(js, exArgs.subject, "B", 20);
-            handler = new ReceiveMessageHandler(true, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 10 total messages, we received: " + messages.size());
 
             // 4. There are still messages left from the last
             // -  Fetch messages, get 10
             System.out.println("----------\n4. Get the rest of the publish.");
-            handler = new ReceiveMessageHandler(true, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 10 total messages, we received: " + messages.size());
 
             // 5. Publish 5 messages
             // -  Fetch messages, get 5
             // -  Since there are less than batch size we only get what the server has.
             System.out.println("----------\n5. Publish 5 which is less than batch size.");
             publish(js, exArgs.subject, "C", 5);
-            handler = new ReceiveMessageHandler(true, 5);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 5 total messages, we received: " + messages.size());
 
             // 6. Publish 15 messages
             // -  Fetch messages, only get 10
             System.out.println("----------\n6. Publish 15 which is more than the batch size.");
             publish(js, exArgs.subject, "D", 15);
-            handler = new ReceiveMessageHandler(true, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 10 total messages, we received: " + messages.size());
 
             // 7. There are 5 messages left
             // -  Fetch messages, only get 5
             System.out.println("----------\n7. There are 5 messages left.");
-            handler = new ReceiveMessageHandler(true, 5);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 5 messages, we received: " + messages.size());
 
             // 8. Read but don't ack.
             // -  Fetch messages, get 10, but either take too long to ack them or don't ack them
             System.out.println("----------\n8. Fetch but don't ack.");
             publish(js, exArgs.subject, "E", 10);
-            handler = new ReceiveMessageHandler(false, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            System.out.println("We should have received 10 message, we received: " + messages.size());
             sleep(3000); // longer than the ackWait
 
             // 9. Fetch messages,
             // -  get the 10 messages we didn't ack
             System.out.println("----------\n9. Fetch, get the messages we did not ack.");
-            handler = new ReceiveMessageHandler(true, 10);
-            sub.receive(10, Duration.ofSeconds(3), handler);
-            handler.latch.await();
+            messages = sub.fetch(10, Duration.ofSeconds(3));
+            report(messages);
+            messages.forEach(Message::ack);
+            System.out.println("We should have received 10 message, we received: " + messages.size());
 
             System.out.println("----------\n");
         }
